+++ src/main/java/com/lazydev/stksongbook/webapp/service/UserSongRatingService.java
@@ -2,7 +2,9 @@ package com.lazydev.stksongbook.webapp.service;
 
 import com.lazydev.stksongbook.webapp.data.model.UserSongRating;
 import com.lazydev.stksongbook.webapp.repository.UserSongRatingRepository;
+import com.lazydev.stksongbook.webapp.security.UserContextService;
 import com.lazydev.stksongbook.webapp.service.exception.EntityNotFoundException;
+import com.lazydev.stksongbook.webapp.service.exception.ForbiddenOperationException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -14,6 +16,7 @@ import java.util.Optional;
 public class UserSongRatingService {
 
   private final UserSongRatingRepository repository;
+  private final UserContextService userContextService;
 
   public List<UserSongRating> findAll() {
     return repository.findAll();
@@ -49,10 +52,16 @@ public class UserSongRatingService {
   }
 
   public UserSongRating save(UserSongRating rating) {
+    if(rating.getUser() != userContextService.getCurrentUser()) {
+      throw new ForbiddenOperationException("No permission.");
+    }
     return repository.save(rating);
   }
 
   public void delete(UserSongRating obj) {
+    if(obj.getUser() != userContextService.getCurrentUser()) {
+      throw new ForbiddenOperationException("No permission.");
+    }
     findByUserIdAndSongId(obj.getUser().getId(), obj.getSong().getId());
     repository.delete(obj);
   }
